Wrap calls in functions
Put a subset of calls supported by alcove into the prx function
namespace. This allows defining types for the calls as well as
intercepting the call to use maps instead of records and do some
portability fixups.

The downside of adding these functions is the increased complexity of
the interface. The functions were manually added to avoid the dreaded
parse transforms and erl_syntax code generation. Any call supported by
alcove that was not included can still be made using the call/3 interface:

    prx:call(Task, mount_define, [ms_private]).