atexit/2: terminate OS process
The second parameter to atexit/2 is a function that allows the caller to
override how the OS process is terminated when the associated Erlang
process exits. atexit/2 is handled by the parent when the child process
exits.

The function takes 3 parameters:
* pid()
* forkchain()
* map() : the result returned by prx:pid/1

Calls must be made directly to the prx_drv process using the
prx_drv:call/4 interface. The default atexit function closes stdio for
the child process in the parent:

        fun(Drv, ForkChain, Pid) ->
            prx_drv:call(Drv, ForkChain, close, [maps:get(stdout, Pid)]),
            prx_drv:call(Drv, ForkChain, close, [maps:get(stdin, Pid)]),
            prx_drv:call(Drv, ForkChain, close, [maps:get(stderr, Pid)])
        end

If the child process ignores closing stdio (for example, the child
process has exec()'ed "sleep 9999999"):

        % send SIGKILL
        Atexit = fun(Drv, ForkChain, Pid) ->
            prx_drv:call(Drv, ForkChain, kill, [maps:get(pid, Pid), 9])
        end,
        prx:atexit(Task, Atexit).

Note:
* the function set by atexit applies to all children of the process

* atexit/2 is not run if the child process has already terminated

  To run code whenever a process exits, the caller should monitor the
  mailbox for exit_status and termsig.