replace_process_image: use fexecve(2)/execve(2)
On FreeBSD and Linux, maintain an open file desciptor to the port
binary and use it to re-exec() the port. For older versions of Linux,
/proc must be mounted, so replace_process_image/1 attempts to fallback
to using the path.

Switch from using execvp(2) to execve(2). replace_process_image/1 will
retrieve the process environment before re-exec()'ing the port.
replace_process_image/3 can be used to set the process environment.