replace_process_image: test for fallback to execve
Add a test on Linux for falling back to execve(2) if /proc is not
available. The test creates a mount namespace and unmounts /proc but
leaves /dev/fd.

A quirk of the fexecve(2) implementation on the system tested is that a
subsequent call to replace_process_image/1 will crash with {error,ebadf}.

The first call to fexecve(2) returns ENOSYS and appears to close the
file descriptor. replace_process_image/1 falls back to using execve(2).

The second call to replace_process_image/1 uses the closed file
descriptor to the prx binary resulting in EBADF when attempting to
remove the close-on-exec flag. Probably EBADF should be ignored for this
fd.