fexecve: set an empty environment for Linux
Unlike fexecve(2) on FreeBSD or execve(2) on Linux, fexecve(2) requires
an environment be specified. Passing in NULL for argv or env will result
in {error,einval}.

Issue exposed by the change in alcove
(84c10fc0e64583d76359223d52f5164bff4737e3) to use NULL to represent the
NIL list in an argv.