task/3,4: transactionally run ops on a process
task/3,4 spawns a new Unix process and runs a sequence of operations. If
the operations succeed, the new task is returned to the caller. If any
of the operations fail, the Unix process is terminated.

Operations are a list of tuples that are run using erlang:apply/3. The
task is prepended to the list of arguments. Success is indicated by
returning 'ok' or {'ok, NewState :: any()} from the function.

task/3,4 is undocumented and experimental for now. The goal is to write
list of instructions for creating sandboxes. For example, an
unprivileged process might use:

* OpenBSD: pledge(2), setrlimit(2)
* Linux: seccomp, setrlimit(2)
* FreeBSD: cap_enter(2), setrlimit(2)

The setrlimit(2) policy can be re-used between OS'es and manipulated by
the caller.