Remove start_child
Fix the reported arity of the function and move from using a function
call to a function body for better error reporting.

start_child was not exported but the idea was to allow an erlang process
such as a supervisor to spawn unix processes on behalf of another erlang
process.

Since any erlang process can request a task to fork if it knows the
erlang pid, not sure if it is useful. The code can still do a:

    gen_fsm:sync_send_event(Task, {start_child, Pid, fork, []},
            infinity).