Handle writes to a non-responsive child process
If a child process does not read stdin, the parent may block waiting
for the pipe buffer to flush.

In exec mode, the parent will send the owning process a message
containing the number of bytes written so far:

    {stdin, pid(), {error, {eagain, non_neg_integer()}}}

The write can be retried later or the process terminated.

In call mode, eagain is fatal since the communication is synchronous. If
the child is not responding, either the the owner is erroneously writing
to the child during a blocking call (e.g., select) or the child is hung
(e.g., due to SIGSTOP).