gen_fsm -> gen_statem
Replace gen_fsm with gen_statem. Suppresses many, many warning messages.

gen_statem swaps the gen_fsm call arguments. The return tuple is also a
wrapped version of the gen_fsm value. These changes are an improvement
and are more consistent with other OTP behaviours. But the conversion
process is a manual one and dialyzer doesn't help much here.

So there are probably bugs. At least the tests are passing.