filter/2: restricts control process calls
Remove calls available to a control process. For example, after forking
a subprocess into a namespace, the control process could disallow
further calls and act as a supervisor for the subprocess.

There is one oddity that should be figured out: when fork/clone are
filtered, undef is returned as an error tuple. For now, explicitly match
{error, undef}.