sh: return if Unix supervisor process crashes
Similar to system(3), sh/2 works by forking an intermediary supervisor
process. If this process exits (for example, due to an unhandled signal
like SIGKILL or SIGCPU), the erlang process calling sh/2 will hang
forever in receive.

Watch for events in the intermediary process and return whatever is
currently in the stdio buffer.