setcpid: child can modify parent process state
Flow control and signaloneof are handled by a prx control process
running in the call state.

If flow control is enabled and the parent control process and child
process are owned by different erlang processes, the child will not be
able to reset the flow control counter.

Allow processes to call setcpid/3 if:
* the calling process is the owner of the control process
* the prx process is a child of the control process