0.16.0
* docs

Use rebar3_ex_doc.

* deps

Update to alcove 0.40.2.

* prx_task

The prx process exits when the statem is stopped and stdin is closed. The
call to exit/2 is redundant. If the prx process is in a bad state
(non-responsive), calling into the process to exit would also hang.

Attempt to signal the process from the parent but ignore errors.

Functions calling with/3 returned `ok | {error, any()}` but with/3
actually can return 2 other tuples: badarg and badop. Since these tuples
aren't part of the type specification, dialyzer complains if code
attempts to match these values.

Remove the `badarg` tuple and rely on the exception on function clause
matching failure.

Explicitly cause an error exception on bad operation and also return the
failing value: a 2-tuple if the return value is not an ok/error tuple
and a 3-tuple for an exception.